Refactors isearch container and reducers
