--- src/components/home/mockData.js
@@ -1,76 +1,0 @@
-export const searchSummary = {
-  city: 'Bodrum',
-  country: 'Turkey',
-  durationInWeeks: 2,
-  paxMix: '2 adults, 2 children',
-  departureDate: 'Sun 13 jul 2016',
-  returnDate: 'Tue 15 jul 2016'
-};
-
-export const filter1 = {
-  part1: 'Are you looking for',
-  bigWord: 'Nightlife',
-  type: 'filter',
-  color: '#81C8BE'
-};
-
-export const filter2 = {
-  part1: 'Are you looking for good',
-  bigWord: 'Food & Drink',
-  type: 'filter',
-  color: '#F39110'
-};
-
-export const filter3 = {
-  part1: 'Are you looking to see the',
-  bigWord: 'Sights',
-  type: 'filter',
-  color: '#8EB8C4'
-};
-
-export const filter4 = {
-  part1: 'Are you looking for',
-  bigWord: 'Sport & Adventure',
-  type: 'filter',
-  color: '#DC3767'
-};
-
-const instagramPrefix = 'https://igcdn-photos-b-a.akamaihd.net/hphotos-ak-xpf1/t51.12885-15/';
-
-export const article1 = {
-  backgroundImage: `${instagramPrefix}s750x750/sh0=108/e35/12479379_141330706245237_420500081_n.jpg`,
-  label: 'Sun and Bathe',
-  title: '10 TURKISH GEMS',
-  overview: 'CLOSE TO BEACH AND CITY',
-  type: 'article'
-};
-
-export const article2 = {
-  backgroundImage: `${instagramPrefix}e35/10472010_1689478414643553_296617682_n.jpg`,
-  label: 'Explore',
-  title: 'TURKEY',
-  overview: '5 HIDDEN SECRETS OF',
-  type: 'article'
-};
-
-export const package1 = {
-  imageSrc: `${instagramPrefix}hphotos-ak-xap1/t51.12885-15/e35/12519323_1674541939492426_1587996260_n.jpg`,
-  caption: 'Child friendly',
-  title: 'Sunwing Side East Beach',
-  subtitle: 'Side, Antalya-area Turkey',
-  numHearts: 4,
-  rating: '4.4/5',
-  packageDetails: 'Breakfast buffet',
-  oldPrice: 30.280,
-  period: '16 nov 2015, 1 week',
-  newPrice: 27.440
-};
-
-export const tiles = [
-  filter1,
-  filter2,
-  article1,
-  filter3,
-  filter4,
-  article2
-];
