--- src/reducers/home.js
@@ -1,30 +1,0 @@
-'use strict';
-
-import { SHOW_ADD_MESSAGE, HIDE_ADD_MESSAGE, QUERY_FETCH_SEARCH_RESULTS } from '../constants/actionTypes';
-
-const initialState = {
-  addMessageVisible: false,
-  loading: true
-};
-
-export default function home (state = initialState, action) {
-  switch (action.type) {
-    case QUERY_FETCH_SEARCH_RESULTS:
-      return ({
-        ...state,
-        loading: false
-      });
-    case SHOW_ADD_MESSAGE:
-      return ({
-        ...state,
-        addMessageVisible: true
-      });
-    case HIDE_ADD_MESSAGE:
-      return ({
-        ...state,
-        addMessageVisible: false
-      });
-    default:
-      return state;
-  }
-}
