+++ src/reducers/utils/mockData.js
@@ -1,0 +1,57 @@
+export const mockTags = [
+  {
+    tagName: 'this',
+    colour: 'red'
+  },
+  {
+    tagName: 'is',
+    colour: 'green'
+  },
+  {
+    tagName: 'sparta',
+    colour: 'pink'
+  }
+];
+
+const instagramPrefix = 'https://igcdn-photos-b-a.akamaihd.net/hphotos-ak-xpf1/t51.12885-15/';
+
+export const mockTiles = [
+  {
+    type: 'article',
+    backgroundImage: `${instagramPrefix}s750x750/sh0=108/e35/12479379_141330706245237_420500081_n.jpg`,
+    label: 'Sun and Bathe',
+    title: '10 TURKISH GEMS',
+    overview: 'CLOSE TO BEACH AND CITY',
+  },
+  {
+    backgroundImage: `${instagramPrefix}e35/10472010_1689478414643553_296617682_n.jpg`,
+    label: 'Explore',
+    title: 'TURKEY',
+    overview: '5 HIDDEN SECRETS OF',
+    type: 'article'
+  },
+  {
+    part1: 'Are you looking for',
+    bigWord: 'Sport & Adventure',
+    type: 'filter',
+    color: '#DC3767'
+  },
+  {
+    part1: 'Are you looking to see the',
+    bigWord: 'Sights',
+    type: 'filter',
+    color: '#8EB8C4'
+  },
+  {
+    part1: 'Are you looking for good',
+    bigWord: 'Food & Drink',
+    type: 'filter',
+    color: '#F39110'
+  },
+  {
+    part1: 'Are you looking for',
+    bigWord: 'Nightlife',
+    type: 'filter',
+    color: '#81C8BE'
+  }
+];
